refactor: move decode_schema alongside encode in params-codec
Symmetric to the encoder lift — same argument applies (smoke runner
asserting on reply payloads, future debug tooling for log inspection).
The decoder is the round-trip counterpart of the encoder and the
existing test suite pins both sides; keeping them in the same crate
avoids a future refactor when someone outside the hub wants to
decode wire bytes back to JSON.

Restructures aether-params-codec into encode + decode submodules
under one re-exporting lib.rs. aether-substrate-hub now re-exports
both pairs (EncodeError + encode_schema, DecodeError + decode_schema)
from the new crate.
