+++ crates/aether-substrate-hub/src/mcp/codecs.rs
@@ -16,9 +16,8 @@ use aether_hub_protocol::{
     EngineId, HubToEngine, KindDescriptor, MailFrame, SchemaType, SessionToken, Uuid,
 };
 
-use aether_params_codec::encode_schema;
+use aether_params_codec::{decode_schema, encode_schema};
 
-use crate::decoder::decode_schema;
 use crate::registry::{EngineRecord, EngineRegistry};
 
 use super::args::MailSpec;
